some example restructuring
